Create game and round function for multiple rounds
I started with the playRound function, which simulates the logic for a
single round of rock paper scissors. The playGame function holds the
entire logic for the game, with the round function being declared within
to keep the implementation clean. The human score and computer score are
also declared and initialised at 0, this will keep track of the score
and be updated throughout the game. A for loop is present after this
which will repeat this five times to simulate five rounds. Finally,
a simple if else section to check the winner overall.